+++ src/common-lib/urls/urls.ts
@@ -282,6 +282,10 @@ type SpecialistSubjectListingLinkProps = {
 
 type WebinarSingleLinkProps = { page: "webinar-single"; webinarSlug: string };
 type BlogSingleLinkProps = { page: "blog-single"; blogSlug: string };
+type CampaignSingleLinkProps = {
+  page: "campaign-single";
+  campaignSlug: string;
+};
 type AboutUsBoardLinkProps = { page: "about-board" };
 type AboutUsWhoWeAreLinkProps = { page: "about-who-we-are" };
 type AboutUsLeadershipLinkProps = { page: "about-leadership" };
@@ -395,6 +399,7 @@ export type OakLinkProps =
   | SpecialistProgrammeListingLinkProps
   | BlogListingLinkProps
   | BlogSingleLinkProps
+  | CampaignSingleLinkProps
   | WebinarListingLinkProps
   | WebinarSingleLinkProps
   | HelpLinkProps
@@ -852,6 +857,12 @@ export const OAK_PAGES: {
     analyticsPageName: "Blog",
     configType: "internal",
     pageType: "blog-single",
+  }),
+  "campaign-single": createOakPageConfig({
+    analyticsPageName: "Campaign",
+    configType: "internal",
+    pageType: "campaign-single",
+    pathPattern: "/campaigns/:campaignSlug",
   }),
   "webinar-single": createOakPageConfig({
     pathPattern: "/webinars/:webinarSlug",