+++ CHANGE_LOG.md
@@ -1,3 +1,75 @@
+# [1.847.0](https://github.com/oaknational/Oak-Web-Application/compare/v1.846.0...v1.847.0) (2025-09-02)
+
+
+### Features
+
+* updates module version ([873ecca](https://github.com/oaknational/Oak-Web-Application/commit/873ecca9a3887efae88b31f037d0b3823a135a3b))
+
+# [1.846.0](https://github.com/oaknational/Oak-Web-Application/compare/v1.845.1...v1.846.0) (2025-09-02)
+
+
+### Bug Fixes
+
+* add missing newline in CampaignPromoBanner stories for better readability ([d337e1b](https://github.com/oaknational/Oak-Web-Application/commit/d337e1b486e068f0e936f8324a264f2c46176742))
+* fix type errors ([f80f10b](https://github.com/oaknational/Oak-Web-Application/commit/f80f10bca82a6927fabe1ce4781be8a4b6f7bde2))
+* increase horizontal padding, justify center on text ([a7082b3](https://github.com/oaknational/Oak-Web-Application/commit/a7082b36c6c7e596c6c45019b39e07758aacf4c6))
+* rm from pupil page ([4dc7e49](https://github.com/oaknational/Oak-Web-Application/commit/4dc7e490ac25b66ad821a9953ddad443f22a64ff))
+* update CampaignPromoBanner stories to include subheading and remove unnecessary parameters ([8949966](https://github.com/oaknational/Oak-Web-Application/commit/89499662067d38d80d12e957f292bc88c4e28d58))
+* update import path for text styles in CampaignPromoBanner component ([65c5055](https://github.com/oaknational/Oak-Web-Application/commit/65c50557ce8a595f1dcbfbf9d0ed756c80f82538))
+
+
+### Features
+
+* add feature flag check to homepage ([8bb9456](https://github.com/oaknational/Oak-Web-Application/commit/8bb945634f3f212b38ad29856ddc0b8899ea4d6a))
+* add text and image to campaigns page ([1eb9541](https://github.com/oaknational/Oak-Web-Application/commit/1eb9541d057b39bcd3ed90e5ae6236455dc3a322))
+* amend homepage sanity query to pull through campaign promo banner ([52f5b1d](https://github.com/oaknational/Oak-Web-Application/commit/52f5b1dc058bfe6d85977dbc15172383871a1363))
+* integrate campaign promo banner into homepage and campaign pages ([c3ef2db](https://github.com/oaknational/Oak-Web-Application/commit/c3ef2dbb9f77b0c5aba2bae2ed4ae949d0f98820))
+* link to campaign from cta ([7df9d8a](https://github.com/oaknational/Oak-Web-Application/commit/7df9d8a129eebc8b9b4f75e03ec30dba66c1fa28))
+* make campaign promo banner component ([5391bbd](https://github.com/oaknational/Oak-Web-Application/commit/5391bbd29ea20cc122d1a4a710dc00ce8a03ad9c))
+
+## [1.845.1](https://github.com/oaknational/Oak-Web-Application/compare/v1.845.0...v1.845.1) (2025-09-02)
+
+# [1.845.0](https://github.com/oaknational/Oak-Web-Application/compare/v1.844.1...v1.845.0) (2025-09-02)
+
+
+### Bug Fixes
+
+* align center on curriculum banner content ([056f62f](https://github.com/oaknational/Oak-Web-Application/commit/056f62f253a4af0caa4df47d3d5c58368ae2fdba))
+* typo ([66ff8ec](https://github.com/oaknational/Oak-Web-Application/commit/66ff8ecc60af46ad0735cc44b624e3c6c32b83e3))
+
+
+### Features
+
+* add banner to subject listing pages ([de3f203](https://github.com/oaknational/Oak-Web-Application/commit/de3f2032d183c43a95e8ecff3064378fe7561885))
+* add banners to required pages ([c116e66](https://github.com/oaknational/Oak-Web-Application/commit/c116e66e20880bd1e242e079623d4e54bc927625))
+* add campaign page type ([1dc7dc7](https://github.com/oaknational/Oak-Web-Application/commit/1dc7dc70564bf69658dcdc7cb7635248d7b7fc06))
+* add link to campaign page in banner ([958ba2e](https://github.com/oaknational/Oak-Web-Application/commit/958ba2e7601f062a72dd2d008ce73c6b95c53fc2))
+* add prop to hide banner when ff disabled ([81aad8c](https://github.com/oaknational/Oak-Web-Application/commit/81aad8c12475d18ab1243923f38a017b924d4bf1))
+* extract PromoSpan to component ([9701d15](https://github.com/oaknational/Oak-Web-Application/commit/9701d1526399cb16a450c6bb63e4391a152fca05))
+* update curriculum banner with optional mythbusting styling ([14ebdd7](https://github.com/oaknational/Oak-Web-Application/commit/14ebdd74d14354e329ab2a022fe7ce14048467dc))
+
+## [1.844.1](https://github.com/oaknational/Oak-Web-Application/compare/v1.844.0...v1.844.1) (2025-08-28)
+
+
+### Bug Fixes
+
+* **a11y:** associate Download button with inline error via aria-describedby ([b1b49cb](https://github.com/oaknational/Oak-Web-Application/commit/b1b49cbcb9a91ce170763a14433176c05d4c87db))
+* **curriculum-downloads:** report download failures via Bugsnag ([ab94faf](https://github.com/oaknational/Oak-Web-Application/commit/ab94faf8944c95fc0fc4c8abe1395011fa5d15fc))
+* **curriculum-downloads:** show inline error on failed download and avoid false success ([a04526d](https://github.com/oaknational/Oak-Web-Application/commit/a04526d7225717a27090c2236d380a1a11741d03))
+
+# [1.844.0](https://github.com/oaknational/Oak-Web-Application/compare/v1.843.0...v1.844.0) (2025-08-28)
+
+
+### Bug Fixes
+
+* spacing between heading on mobile ([3393042](https://github.com/oaknational/Oak-Web-Application/commit/3393042bb198c0ab20368ff45da9a77065825aec))
+* unique key for component ([49d95a2](https://github.com/oaknational/Oak-Web-Application/commit/49d95a2b956018722d1118fe04fd0699ebecccec))
+
+
+### Features
+
+* add campaign intro component ([7d2a290](https://github.com/oaknational/Oak-Web-Application/commit/7d2a290217a53941eef335ed8862fef0b5a01451))
+
 # [1.843.0](https://github.com/oaknational/Oak-Web-Application/compare/v1.842.1...v1.843.0) (2025-08-27)